neetmentor/requirements.txt
Django
djangorestframework
django-cors-headers
whitenoise
gunicorn
dj-database-url
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
